% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concordance.test.R
\docType{methods}
\name{coef}
\alias{coef}
\alias{coef,ConcordanceTest-method}
\title{Extract test results from the results of a concordance.test}
\usage{
\S4method{coef}{ConcordanceTest}(object, ...)
}
\arguments{
\item{object}{An object of \code{\link{ConcordanceTest-class}}}

\item{...}{Not used}
}
\value{
A matrix
}
\description{
\code{coef} extract the test results from the results of a concordance.test
}
\examples{

matRandom <- matrix(rnorm(3*20),20,3)
testResult <- concordance.test(matRandom)
getPsi(testResult)
coef(testResult)

}
\seealso{
Other concordance functions: \code{\link{concordance.test}},
  \code{\link{getPsi}}, \code{\link{rfromPsi}}
}
