% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concordance.test.R
\docType{class}
\name{ConcordanceTest-class}
\alias{ConcordanceTest-class}
\title{Class ConcordanceTest}
\description{
This class stores results obtained from a concordance test.
}
\details{
Class \code{ConcordanceTest} stores results from a concordance test.
}
\section{Slots}{

\describe{
\item{\code{pvalue}}{The pvalue}

\item{\code{psi1}}{The concordance in matrix \code{x}}

\item{\code{psi2}}{The concordance in matrix \code{y}}

\item{\code{method}}{The method used to obtain the pvalue}

\item{\code{alternative}}{The alternative hypothesis}

\item{\code{ci.lower}}{The lower confidence boudary}

\item{\code{ci.upper}}{The upper confidence boudary}

\item{\code{ci.method}}{The method used to obtain the confidence interval}

\item{\code{alpha}}{The significance level}

\item{\code{call}}{The call made to the \code{\link{concordance.test}} function}
}}

