% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concordance.test.R
\docType{methods}
\name{$,ConcordanceTest-method}
\alias{$,ConcordanceTest-method}
\alias{get,ConcordanceTest-method}
\title{Extract argument values from a ConcordanceTest object}
\usage{
\S4method{$}{ConcordanceTest}(x, name)
}
\arguments{
\item{x}{An object of \code{\link{ConcordanceTest-class}}}

\item{name}{The argument to get the value of}
}
\value{
The value of the requested argument
}
\description{
Extracts argument values from a \code{\link{ConcordanceTest-class}} object
}
\examples{

matRandom <- matrix(rnorm(3*20),20,3)
testResult <- concordance.test(matRandom)
names(testResult)
testResult$psi
}
