% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concordance.test.R
\docType{methods}
\name{names,ConcordanceTest-method}
\alias{names,ConcordanceTest-method}
\title{Extract argument names from a ConcordanceTest object}
\usage{
\S4method{names}{ConcordanceTest}(x)
}
\arguments{
\item{x}{An object of \code{\link{ConcordanceTest-class}}}
}
\value{
A character vector
}
\description{
\code{names} extracts argument names from a \code{\link{ConcordanceTest-class}} object
}
\examples{

matRandom <- matrix(rnorm(3*20),20,3)
testResult <- concordance.test(matRandom)
names(testResult)
}
