% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{scores}
\alias{scores}
\title{Hypothetical data of outcomes for two risk models}
\format{A list with two elements named 'modelA' and 'modelB' both containing a 
dataframe with outcome scores for 100 subjects in the rows each having two 
replicate measurements in the columns.}
\source{
\code{vignette("nopaco", package = "nopaco")}
}
\usage{
data(scores)
}
\description{
This data is generated as explained in the nopaco vignette. 
It represents the outcomes of the risk models (model A and model B). 
Both models were applied to gene expression profiles 100 subjects, each 
run in duplo.
}
\examples{
data(scores)
str(scores)
plot(scores[['modelA']])
plot(scores[['modelB']])
}
\keyword{datasets}
